+++ precompile/precompileconfig/registerer.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"sort"
 
-	"github.com/ava-labs/subnet-evm/constants"
-	"github.com/ava-labs/subnet-evm/utils"
 	"github.com/ethereum/go-ethereum/common"
 )
 
@@ -13,45 +11,12 @@ var (
 	// registeredModules is a list of Module to preserve order
 	// for deterministic iteration
 	registeredModules = make([]Module, 0)
-
-	reservedRanges = []utils.AddressRange{
-		{
-			Start: common.HexToAddress("0x0100000000000000000000000000000000000000"),
-			End:   common.HexToAddress("0x01000000000000000000000000000000000000ff"),
-		},
-		{
-			Start: common.HexToAddress("0x0200000000000000000000000000000000000000"),
-			End:   common.HexToAddress("0x02000000000000000000000000000000000000ff"),
-		},
-		{
-			Start: common.HexToAddress("0x0300000000000000000000000000000000000000"),
-			End:   common.HexToAddress("0x03000000000000000000000000000000000000ff"),
-		},
-	}
 )
-
-// ReservedAddress returns true if [addr] is in a reserved range for custom precompiles
-func ReservedAddress(addr common.Address) bool {
-	for _, reservedRange := range reservedRanges {
-		if reservedRange.Contains(addr) {
-			return true
-		}
-	}
-
-	return false
-}
 
 // RegisterModule registers a stateful precompile module
 func RegisterModule(stm Module) error {
 	address := stm.Address
 	key := stm.ConfigKey
-
-	if address == constants.BlackholeAddr {
-		return fmt.Errorf("address %s overlaps with blackhole address", address)
-	}
-	if !ReservedAddress(address) {
-		return fmt.Errorf("address %s not in a reserved range", address)
-	}
 
 	for _, registeredModule := range registeredModules {
 		if registeredModule.ConfigKey == key {
